+++ src/variables/SeomaticVariable.php
@@ -26,7 +26,8 @@ use nystudio107\seomatic\services\Tag;
 use nystudio107\seomatic\services\Title;
 use nystudio107\seomatic\services\MetaContainers;
 use nystudio107\seomatic\services\MetaBundles;
-use nystudio107\seomatic\variables\ManifestVariable as Manifest;
+
+use nystudio107\pluginmanifest\variables\ManifestVariable as Manifest;
 
 use Craft;
 
@@ -98,7 +99,10 @@ class SeomaticVariable extends ServiceLocator
             'script' => Script::class,
             'tag' => Tag::class,
             'title' => Title::class,
-            'manifest' => Manifest::class
+            'manifest' => [
+                'class' => Manifest::class,
+                'manifestService' => Seomatic::$plugin->manifest,
+            ]
         ];
 
         $config['components'] = $components;